Add cygwin::enable parameter
It's occasionally useful to use `cygwin::windows_path()` on non-Cygwin
nodes, e.g. for exported resources. That can now be done by setting
`enable` to `false`.

Additionally, a few of the modules I've been working on need to know if
Cygwin will be used before it is installed (on the first run). For
example, [ploperations/ssh][] needs to know if it should installed
Cygwin OpenSSH or native OpenSSH.

[ploperations/ssh]: https://github.com/ploperations/ploperations-ssh